Fixed Sliding Items
The problem was that the dock island found its velocity parallel to the
rail, moved accordingly, check if it went too far, and, if true, set
location to the end of the rail. The item wouldn’t move at all if this
happened, and the two would separate a little. So basically, the dock
island teleported and its velocity (which the item relies on) didn’t
reflect that.